% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cgce_df}
\alias{cgce_df}
\title{Comercio Brasileiro por CGCE}
\format{
o banco de dados possui cinco variáveis: \code{path},
que indica a direção do comércio (exportações ou importações),
\code{co_cgce_n1}, que indica o código do cgce,
\code{co_ano}, \code{co_pais} e \code{value}.
}
\source{
\url{https://www.gov.br/produtividade-e-comercio-exterior/pt-br/assuntos/comercio-exterior/estatisticas/base-de-dados-bruta}
}
\usage{
cgce_df
}
\description{
Dados de comercio bilateral oferecidos pelo Ministério da Economia
agregados pela classificação de Grandes Categorias Econômicas.
}
\details{
Classificação utilizada pelo IBGE em seu sistema de contas nacionais. Dividi-se em
Bens de Capital (BK), Bens Intermediários (BI), Bens de Consumo (BC) e Bens Não Especificados
Anteriormente. A referência de cada um dos produtos pode ser encontrada na
base \code{dic_ncm_cgce}, disponível neste pacote.

Nota Metodológica do Ministério da Economia: https://balanca.economia.gov.br/balanca/metodologia/Nota_CGCE.pdf
}
\keyword{datasets}
