% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cuci_df}
\alias{cuci_df}
\title{Comercio Brasileiro por CUCI}
\format{
o banco de dados possui cinco variáveis: \code{path},
que indica a direção do comércio (exportações ou importações),
\code{co_cuci_sec}, que indica o código cuci,
\code{co_ano}, \code{co_pais} e \code{value}.
}
\source{
\url{https://www.gov.br/produtividade-e-comercio-exterior/pt-br/assuntos/comercio-exterior/estatisticas/base-de-dados-bruta}
}
\usage{
cuci_df
}
\description{
Dados de comercio bilateral oferecidos pelo Ministério da Economia
agregados pela Classificação Uniforme para o Comércio Internacional (CUCI).
}
\details{
Classificação criada pela ONU para estatísticas de comércio exterior. Segundo o
Ministério da Economia, os agrupamentos da CUCI refletem: materiais utilizados na produção;
estágio de processamento; práticas de mercado; importância dos bens e mudanças tecnológicas.
}
\keyword{datasets}
