% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dic_blocos}
\alias{dic_blocos}
\title{Lista de Blocos de Países}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 324 rows and 2 columns.
}
\usage{
dic_blocos
}
\description{
Correlaciona países a blocos.
}
\keyword{datasets}
