% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dic_ncm_isic}
\alias{dic_ncm_isic}
\title{Dicionário de Correlação NCM - ISIC}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13118 rows and 3 columns.
}
\usage{
dic_ncm_isic
}
\description{
Correlaciona produtos em NCM ao sistema de classificação ISIC.
}
\keyword{datasets}
