% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dic_paises}
\alias{dic_paises}
\title{Dicionário de Correlação Códigos Países - Nomes de Países}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 281 rows and 2 columns.
}
\usage{
dic_paises
}
\description{
Correlaciona países, seguindo os códigos oferecidos pelo Ministério da Economia,
com os nomes na grafia estabelecidas pelo Ministério da Economia.
}
\keyword{datasets}
