% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fator_df}
\alias{fator_df}
\title{Comercio Brasileiro por Fator Agregado}
\format{
o banco de dados possui cinco variáveis: \code{path},
que indica a direção do comércio (exportações ou importações),
\code{co_fat_agreg}, que indica o código do fator agregado,
\code{co_ano}, \code{co_pais} e \code{value}.
}
\source{
\url{https://www.gov.br/produtividade-e-comercio-exterior/pt-br/assuntos/comercio-exterior/estatisticas/base-de-dados-bruta}
}
\usage{
fator_df
}
\description{
Dados de comercio bilateral oferecidos pelo Ministério da Economia
agregados pela classificação de Fator Agregado. Dados originais podem ser acessados em
https://www.gov.br/produtividade-e-comercio-exterior/pt-br/assuntos/comercio-exterior/estatisticas/base-de-dados-bruta
}
\details{
Essa classificação foi desenvolvida pelo Banco do Brasil, na década de 70 e,
atualmente, deixou de ser disponibilizada pelo Comex Stat. A classificação divide
os produtos em três níveis: básicos, semimanufaturados e manufaturados. Os dois últimos
são costumam ser chamados de "produtos industrializados". A referência de cada um
dos produtos pode ser encontrada na base \code{dic_ncm_fator}, disponível neste pacote.
}
\keyword{datasets}
