% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isic_df}
\alias{isic_df}
\title{Comercio Brasileiro por ISIC}
\format{
o banco de dados possui cinco variáveis: \code{path},
que indica a direção do comércio (exportações ou importações),
\code{co_isic_secao}, que indica o código isic,
\code{co_ano}, \code{co_pais} e \code{value}.
}
\source{
\url{https://www.gov.br/produtividade-e-comercio-exterior/pt-br/assuntos/comercio-exterior/estatisticas/base-de-dados-bruta}
}
\usage{
isic_df
}
\description{
Dados de comercio bilateral oferecidos pelo Ministério da Economia
agregados pela Standard Industrial Classification of All Economic Activities (ISIC).
}
\details{
Classificação criada pela ONU para estatísticas de comércio exterior. Propõe estrutura
menos voltada para produto e mais voltada para atividade produtiva. Mais informações:
https://balanca.economia.gov.br/balanca/metodologia/Nota_ISIC-CUCI.pdf
}
\keyword{datasets}
