% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sh4_df}
\alias{sh4_df}
\title{Comercio Brasileiro em SH4}
\format{
o banco de dados possui sete variáveis: \code{path},
que indica a direção do comércio (exportações ou importações),
\code{co_ano}, \code{co_mes}, \code{co_pais}, o código do país,
segundo a base de dados do MEcon, \code{co_ncm_secrom}, que faz
referência ao nível de agregação 4 do Sistema Harmonizado de Comércio,
 \code{value} o montante comercializado e \code{no_pais}.
}
\source{
\url{https://www.gov.br/produtividade-e-comercio-exterior/pt-br/assuntos/comercio-exterior/estatisticas/base-de-dados-bruta}
}
\usage{
sh4_df
}
\description{
Dados de comercio bilateral oferecidos pelo Ministério da Economia
em nível de agregação SH4.
}
\keyword{datasets}
