% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dic_sh6_sh2}
\alias{dic_sh6_sh2}
\title{Dicionário de Correlação SH6 - SH1}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 6308 rows and 3 columns.
}
\usage{
dic_sh6_sh2
}
\description{
Correlaciona produtos do Sistema Harmonizado em SH6 para SH2.
}
\keyword{datasets}
