% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dic_sh6_sh4}
\alias{dic_sh6_sh4}
\title{Dicionário de Correlação SH6 - SH4}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1287 rows and 2 columns.
}
\usage{
dic_sh6_sh4
}
\description{
Correlaciona produtos do Sistema Harmonizado em SH6 para SH4.
}
\keyword{datasets}
